+++ ImageAlpha-Bridging-Header.h
@@ -1,1 +1,2 @@
 #import "libimagequant.h"
+#include <zlib.h>